seperate fullstory suffixed props to a new class
